script

+++ src/BasicNft.sol
@@ -4,9 +4,8 @@ pragma solidity 0.8.21;
 import {ERC721} from "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 
 contract BasicNft is ERC721 {
-
-    uint private s_tokenCounter;
-    mapping(uint => string) private s_tokenIdToUri;
+    uint256 private s_tokenCounter;
+    mapping(uint256 => string) private s_tokenIdToUri;
 
     constructor() ERC721("The Dawgs", "DOG") {
         s_tokenCounter = 0;
@@ -18,7 +17,7 @@ contract BasicNft is ERC721 {
         s_tokenCounter++;
     }
 
-    function tokenURI(uint tokenId) public view override returns(string memory) {
+    function tokenURI(uint256 tokenId) public view override returns (string memory) {
         return s_tokenIdToUri[tokenId];
     }
 }
